requirements.txt
bing_image_downloader==1.1.2
boto3==1.35.14
botocore==1.35.14
kafka==1.3.5
pandas==2.2.3
Pillow==11.0.0
psycopg==3.2.3
streamlit==1.39.0
